requirements.txt
pandas
pyjanitor
matplotlib
seaborn
runpynb
pranaam